+++ src/App.vue
@@ -6,7 +6,26 @@
 
 <script>
 export default {
-  name: 'App'
+  name: 'App',
+  data () {
+    return {
+      user: {}
+    }
+  },
+  mounted () {
+    this.init()
+  },
+  methods: {
+    init () {
+      this.$firebase.auth.onAuthStateChanged((user) => {
+        if (!user) {
+          console.log('no user')
+        } else {
+          this.$bindAsObject('user', this.$firebase.user(user.uid))
+        }
+      })
+    }
+  }
 }
 </script>