+++ src/layouts/builder.vue
@@ -50,9 +50,13 @@
     >
       <!-- QScrollArea is optional -->
       <q-scroll-area class="fit q-pa-sm">
-        <q-item to="/odashboard">
+        <q-item to="/dashboard">
           <q-item-tile icon="fa-home" />
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Dashboard</q-tooltip>
+        </q-item>
+        <q-item to="/realdashboard" v-if="$root.user.realUser">
+          <q-item-tile image><img src="statics/logo_square@0.5x.png" style="width: 30px;" /></q-item-tile>
+          <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">REAL Curriculum</q-tooltip>
         </q-item>
         <q-item-separator />
         <q-item :to="{ name: 'olist', params: { type: 'sermons' }}">
@@ -64,11 +68,36 @@
           <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Lessons</q-tooltip>
         </q-item>
         <q-item-separator />
+        <div v-if="!$root.user.nqUser">
+          <q-item :to="{ name: 'olist', params: { type: 'quotes' }}">
+            <q-item-tile icon="fa-quote-left" />
+            <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Quotes</q-tooltip>
+          </q-item>
+          <q-item :to="{ name: 'olist', params: { type: 'images' }}">
+            <q-item-tile icon="fa-image" />
+            <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Images</q-tooltip>
+          </q-item>
+          <q-item :to="{ name: 'olist', params: { type: 'illustrations' }}">
+            <q-item-tile icon="fa-paint-brush" />
+            <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Illustrations</q-tooltip>
+          </q-item>
+          <q-item :to="{ name: 'olist', params: { type: 'lyrics' }}">
+            <q-item-tile icon="fa-music" />
+            <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Lyrics</q-tooltip>
+          </q-item>
+          <q-item :to="{ name: 'olist', params: { type: 'videos' }}">
+            <q-item-tile icon="fa-play" />
+            <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Videos</q-tooltip>
+          </q-item>
+        </div>
+        <q-item-separator v-if="!$root.user.nqUser" />
         <q-item>
           <q-item-tile icon="fa-cog" />
+          <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Settings</q-tooltip>
         </q-item>
         <q-item link @click.native="logout">
           <q-item-tile icon="fa-sign-out-alt fab" />
+          <q-tooltip anchor="center left" self="center right" :disable="$q.platform.is.mobile">Log Out</q-tooltip>
         </q-item>
       </q-scroll-area>
     </q-layout-drawer>
@@ -103,7 +132,7 @@ export default {
   data () {
     return {
       leftDrawer: this.$route.name !== 'olesson',
-      rightDrawer: this.$route.name === 'olesson' && this.$q.platform.is.desktop,
+      rightDrawer: this.$route.name === 'olesson' && this.$q.platform.is.desktop && this.$root.user.nqUser,
       showRightDrawer: this.$route.name === 'olesson' && this.$q.platform.is.desktop,
       pageType: this.$route.name
     }
@@ -113,7 +142,7 @@ export default {
       console.log('page change')
       this.pageType = val
       this.leftDrawer = val !== 'olesson'
-      this.rightDrawer = val === 'olesson' && this.$q.platform.is.desktop
+      this.rightDrawer = val === 'olesson' && this.$q.platform.is.desktop && this.$root.user.nqUser
       this.showRightDrawer = val === 'olesson' && this.$q.platform.is.desktop
     }
   },