+++ src/router/routes.js
@@ -6,7 +6,7 @@ export default [
     path: '/',
     component: () => import('layouts/builder'),
     children: [
-      { path: 'odashboard', name: 'odashboard', component: () => import('pages/ODashboard') },
+      { path: 'dashboard', name: 'dashboard', component: () => import('pages/ODashboard') },
       { path: 'olist/:type', name: 'olist', component: () => import('pages/OList') },
       { path: 'oseries/:id', name: 'oseries', component: () => import('pages/OSeries') },
       { path: 'olesson/:id', name: 'olesson', component: () => import('pages/OLesson') },
@@ -22,7 +22,7 @@ export default [
     path: '/',
     component: () => import('layouts/real'),
     children: [
-      { path: 'rdashboard', name: 'rdashboard', component: () => import('pages/RDashboard') },
+      { path: 'realdashboard', name: 'realdashboard', component: () => import('pages/RDashboard') },
       { path: 'rlist/:type', name: 'rlist', component: () => import('pages/RList') },
       { path: 'rseries/:id', name: 'rseries', component: () => import('pages/RSeries') }
     ],